chore: move create_job_applicant to test_utils to prevent circular import
(cherry picked from commit dee1819f7f359e31e651dc69cfe8b2726ce7de2a)
