+++ hrms/tests/test_utils.py
@@ -2,6 +2,7 @@ import frappe
 from frappe.utils import add_months, get_first_day, get_last_day, getdate, now_datetime
 
 from erpnext.setup.doctype.department.department import get_abbreviated_name
+from erpnext.setup.doctype.designation.test_designation import create_designation
 from erpnext.setup.utils import enable_all_roles_and_domains
 
 
@@ -92,3 +93,25 @@ def create_department(name: str, company: str = "_Test Company") -> str:
 	department.update({"doctype": "Department", "department_name": name, "company": "_Test Company"})
 	department.insert()
 	return department.name
+
+
+def create_job_applicant(**args):
+	args = frappe._dict(args)
+	filters = {
+		"applicant_name": args.applicant_name or "_Test Applicant",
+		"email_id": args.email_id or "test_applicant@example.com",
+	}
+
+	if frappe.db.exists("Job Applicant", filters):
+		return frappe.get_doc("Job Applicant", filters)
+
+	job_applicant = frappe.get_doc(
+		{
+			"doctype": "Job Applicant",
+			"status": args.status or "Open",
+			"designation": create_designation().name,
+		}
+	)
+	job_applicant.update(filters)
+	job_applicant.save()
+	return job_applicant
